+++ test_int_job_process_sleepy_conversations.py
@@ -3,8 +3,8 @@ import random
 import string
 import datetime
 
-from sleepy_conversations import get_sleepy_conversations,change_conversation_state,get_conversation_state, send_message, wakeup_sleepy_conversation, is_last_message_from_operator, job_process_sleepy_conversations
-from utils import get_messages
+from src.job_process_sleepy_conversations import get_sleepy_conversations,change_conversation_state,get_conversation_state, send_message, wakeup_sleepy_conversation, is_last_message_from_operator, job_process_sleepy_conversations
+from src.utils import get_messages
 
 conversation_id = "session_691ea0d5-0543-427b-85dd-95bc412ceb27"
